Started work on a pure css decorator
This mainly as an example on how to do a decorator since I don't use
pure css. For the moment only supports standard input and textarea
and has trouble with descriptions. Doesn't seem to be any standard pure
css support for them when using an aligned form.

Also no validation feedback (there is none in pure css)